deep-learning/projects/traffic_sign_detection/requirements.txt
tensorflow == 2.3.1
numpy
opencv-python
time
os
matplotlib	
pandas
pillow
glob
random
pickle
